% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFunctions.R
\name{add.pw}
\alias{add.pw}
\title{Add a password}
\usage{
add.pw(key, site, id, pw)
}
\arguments{
\item{key}{Numeric key.}

\item{site}{Character string. Website name.}

\item{id}{Character string. Site ID.}

\item{pw}{Character String. Password.}
}
\description{
Add a new passowrd to the passwords dataframe.
}
\examples{
add.pw(key, site, id, pw)
}
\keyword{password}

