% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChickenTime.R
\name{Chicken.Time}
\alias{Chicken.Time}
\title{Make a Chicken Dinner}
\usage{
Chicken.Time(weight, end.time, smart = TRUE)
}
\arguments{
\item{weight}{The weight of your chicken in kilograms. No default.}

\item{end.time}{Character string. The time you want your dinner to be ready in the format "hh:mm". If not set will asume asap.}

\item{smart}{TRUE or FALSE. When set to TRUE will return an error if the requested time set in end.time is not possible. Defaults to TRUE.}
}
\description{
This function allows you to make a perfect chicken dinner.
}
\examples{
Chicken.Time(1.35, "18:55", TRUE)
}
\keyword{chicken}
\keyword{time}

