% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFunctions.R
\name{lock.it}
\alias{lock.it}
\title{Password lock}
\usage{
lock.it(key, pwd)
}
\arguments{
\item{key}{Numeric key.}

\item{pwd}{dataframe}
}
\description{
Lock the passwords dataframe.
}
\examples{
lock.it(key, pwd)
}
\keyword{password}

