% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFunctions.R
\name{makeCode}
\alias{makeCode}
\title{Make password code}
\usage{
makeCode(key)
}
\arguments{
\item{key}{Numeric key.}

\item{pwd}{dataframe}
}
\description{
Create the code from the key.
}
\examples{
makeCode(key)
}
\keyword{password}

