% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefreshPackage.R
\name{refresh.wil}
\alias{refresh.wil}
\title{Re-install the wil package from GitHub}
\usage{
refresh.wil(loc)
}
\arguments{
\item{loc}{Either home or work}
}
\description{
This function downloads the wil package from GitHub and installs it.
}
\keyword{GitHub}
\keyword{wil}

