% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFunctions.R
\name{unlock.it}
\alias{unlock.it}
\title{Password unlock}
\usage{
unlock.it(key)
}
\arguments{
\item{key}{Numeric key.}
}
\description{
unlock the passwords dataframe.
}
\examples{
unlock.it(key)
}
\keyword{password}

